requirements.txt
torch>=2.0.0
transformers>=4.36.0
datasets>=2.18.0
accelerate>=0.26.0
peft>=0.8.0
evaluate>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.0
scikit-learn>=1.3.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.15.0
sentencepiece>=0.1.99
sacrebleu>=2.3.0
bert-score>=0.3.13
